Move utilities to utils folder.
